Cleaning in progress...
git-svn-id: file:///tmp/svnroot_only_rec/ReC/trunk@31 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ src/com/linkare/rec/acquisition/MaxPacketNumUnknown.java
@@ -2,5 +2,5 @@ package com.linkare.rec.acquisition;
 
 public interface MaxPacketNumUnknown
 {
-	public static final int value = (int)((int)-1);
+	public static final int value = -1;
 }
